feat: add modules part of task
